refactor: removed unnecessary folders
